Checkbox: add initial implementation
checked property is unused as it seems to be not used in PF examples
is_checked is implemented as an Option<Option<bool>> to differentiate between the default case where the is_checked prop is not set and the null case (indeterminate).  The parent component should always set the is_checked property as Some(Option<bool>) and leave the base None case for the property being default.